Re-enable the flake8 pre-commit checker.
It was failing before because of use of 'long' integers. There isn't
really a distinction between 'int' and 'long' any more so just use
ints in the code to fix the problem.

+++ dockalot/config.py
@@ -31,7 +31,7 @@ def string_importer(value, name):
 
 def integer_importer(value, name):
     try:
-        return long(value)
+        return int(value)
     except:
         raise ConfigurationError("Configuration value '{}' is not an integer"
             .format('.'.join(name)))
